% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_E_random.R
\name{build_E_random}
\alias{build_E_random}
\title{Generate a list of introduction events}
\usage{
build_E_random(
  period_start,
  period_end = period_start,
  n_events = NULL,
  n_ind = NULL,
  stage = "Eh",
  loc
)
}
\arguments{
\item{period_start}{Date in '\%Y-\%m-\%d' format. Defines the beginning of the period for introduction.}

\item{period_end}{Date in '\%Y-\%m-\%d' format. Defines the end of the period for introduction.}

\item{n_events}{Numeric value. Number of introduction events generated.}

\item{n_ind}{Numeric value. Number of individuals introduced.}

\item{stage}{Character vector. Epidemiological stage of introduced individuals.}

\item{loc}{Character or numeric vector. List of patches to seed random introductions.}
}
\value{
Data frame with scheduled introductions.
}
\description{
Function used to generate a random list of schedule introduction.
The introduction events generated can occur in different locations and at different dates in a given period.
The number of individuals introduced is randomly distributed over all the events.
If the number of events is not provided, location and date are randomly sampled for each introduced individual.
The epidemiological stage can be either deterministic (unique value provided) or random (randomly selected in a vector of values).
}
\examples{
\dontrun{
build_E_random(period_start = as.Date("2020/03/10"),
period_end = as.Date("2020/09/30"),
n_ind = NULL, n_events = 10,
stage = "Eh", loc = LETTERS[1:3])
}

}
\keyword{events}
