% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gdata.R
\name{build_gdata}
\alias{build_gdata}
\title{Generate global parameters list}
\usage{
build_gdata(
  vector = "Ae. albopictus (D)",
  virus = "DEN",
  bHM = NULL,
  bMH = NULL,
  muH = NULL,
  rhoH = NULL,
  muE = NULL,
  TE = NULL,
  TDDE = NULL,
  q1E = NULL,
  q2E = NULL,
  q3E = NULL,
  q4E = NULL,
  q5E = NULL,
  q6E = NULL,
  mu1L = NULL,
  mu2L = NULL,
  mu3L = NULL,
  q1L = NULL,
  q2L = NULL,
  q3L = NULL,
  q4L = NULL,
  q5L = NULL,
  q6L = NULL,
  muEM = NULL,
  mu1P = NULL,
  mu2P = NULL,
  mu3P = NULL,
  q1P = NULL,
  q2P = NULL,
  q3P = NULL,
  q4P = NULL,
  q5P = NULL,
  q6P = NULL,
  mu1A = NULL,
  mu2A = NULL,
  mu3A = NULL,
  gammaAem = NULL,
  sigma = NULL,
  gammaAh = NULL,
  muR = NULL,
  TAG = NULL,
  TDDAG = NULL,
  q1Ag = NULL,
  q2Ag = NULL,
  q3Ag = NULL,
  q4Ag = NULL,
  q5Ag = NULL,
  q6Ag = NULL,
  gammaAo = NULL,
  beta1 = NULL,
  beta2 = NULL,
  maxbite = NULL,
  startFav = NULL,
  endFav = NULL,
  muErain = NULL,
  muLrain = NULL,
  muPrain = NULL,
  verbose = T
)
}
\arguments{
\item{vector}{string. "Ae. albopictus", "Ae. albopictus (D)" or "Ae. aegypti". Default is "Ae. albopictus (D)".}

\item{virus}{string.  "DEN" (dengue), "ZIK" (zika) or "CHI" (chikungunya). Default is "DEN" (dengue) virus.}

\item{bHM}{numeric. Probability of infection from host to vector when an infected host is bitten by an susceptible vector}

\item{bMH}{numeric. Probability of infection from vector to host when a human is bitten by an infected mosquito.}

\item{muH}{numeric. Daily transition rate from exposed (E) to infected (I) for humans (1/days)}

\item{rhoH}{numeric. Daily recovery rate of humans (1/days)}

\item{muE}{numeric. Daily mortality rate of eggs (1/days)}

\item{TE}{numeric. Minimal temperature needed for albopictus eggs development (째C)  (see details of albopictus functions)}

\item{TDDE}{numeric. Total number of degree-day necessary for albopictus eggs development (째C) (see details of albopictus functions)}

\item{q1E}{numeric. Parameter for the aegypti egg development function (see details of aegypti functions)}

\item{q2E}{numeric. Parameter for the aegypti egg development function (see details of aegypti functions)}

\item{q3E}{numeric. Parameter for the aegypti egg development function (see details of aegypti functions)}

\item{q4E}{numeric. Parameter for the aegypti egg development function (see details of aegypti functions)}

\item{q5E}{numeric. Parameter for the aegypti egg development function (see details of aegypti functions)}

\item{q6E}{numeric. Parameter for the aegypti egg development function (see details of aegypti functions)}

\item{mu1L}{numeric. Parameter for the larvae mortality function (see details)}

\item{mu2L}{numeric. Parameter for the larvae mortality function (see details)}

\item{mu3L}{numeric. Parameter for the larvae mortality function (see details)}

\item{q1L}{numeric. Parameter for the function of transition from larva to pupa (see details)}

\item{q2L}{numeric. Parameter for the function of transition from larva to pupa (see details)}

\item{q3L}{numeric. Parameter for the function of transition from larva to pupa (see details)}

\item{q4L}{numeric. Parameter for the function of transition from larva to pupa (see details of aegypti functions)}

\item{q5L}{numeric. Parameter for the function of transition from larva to pupa (see details of aegypti functions)}

\item{q6L}{numeric. Parameter for the function of transition from larva to pupa (see details of aegypti functions)}

\item{muEM}{numeric. Daily mortality rate of emerging adults during the emergence (1/days)}

\item{mu1P}{numeric. Parameter for the pupae mortality function (see details)}

\item{mu2P}{numeric. Parameter for the pupae mortality function (see details)}

\item{mu3P}{numeric. Parameter for the pupae mortality function (see details)}

\item{q1P}{numeric. Parameter for the function of transition from pupae to emerging adult (see details)}

\item{q2P}{numeric. Parameter for the function of transition from pupae to emerging adult (see details)}

\item{q3P}{numeric. Parameter for the function of transition from pupae to emerging adult (see details)}

\item{q4P}{numeric. Parameter for the function of transition from pupae to emerging adult (see details of aegypti functions)}

\item{q5P}{numeric. Parameter for the function of transition from pupae to emerging adult (see details of aegypti functions)}

\item{q6P}{numeric. Parameter for the function of transition from pupae to emerging adult (see details of aegypti functions)}

\item{mu1A}{numeric. Parameter for the adult mortality function: \eqn{mu1A \times e^{(temperature_{t}-10) \times  mu2A}  + mu3A}}

\item{mu2A}{numeric. Parameter for the adult mortality function: \eqn{mu1A \times e^{(temperature_{t}-10) \times  mu2A}  + mu3A}}

\item{mu3A}{numeric. Parameter for the adult mortality function: \eqn{mu1A \times e^{(temperature_{t}-10) \times  mu2A}  + mu3A}}

\item{gammaAem}{numeric. Daily development rate of emerging adults (1/days)}

\item{sigma}{numeric. Sex-ratio at the emergence}

\item{gammaAh}{numeric. Daily transition rate from host-seeking to engorged adults (1/days)}

\item{muR}{numeric. Daily mortality rate related to seeking behavior (1/days)}

\item{TAG}{numeric. Minimal temperature needed for egg maturation (째C) (see details of albopictus functions)}

\item{TDDAG}{numeric. Total number of degree-days necessary for egg maturation (째C) (see details of albopictus functions)}

\item{q1Ag}{numeric. Parameter for the function driving eggs maturation after blood meal (see details of aegypti functions)}

\item{q2Ag}{numeric. Parameter for the function driving eggs maturation after blood meal (see details of aegypti functions)}

\item{q3Ag}{numeric. Parameter for the function driving eggs maturation after blood meal (see details of aegypti functions)}

\item{q4Ag}{numeric. Parameter for the function driving eggs maturation after blood meal (see details of aegypti functions)}

\item{q5Ag}{numeric. Parameter for the function driving eggs maturation after blood meal (see details of aegypti functions)}

\item{q6Ag}{numeric. Parameter for the function driving eggs maturation after blood meal (see details of aegypti functions)}

\item{gammaAo}{numeric. Daily transition rate from oviposition site-seeking to host-seeking adults (1/days)}

\item{beta1}{numeric.  Number of eggs laid by ovipositing nulliparous females (per female)}

\item{beta2}{numeric.  Number of eggs laid by ovipositing parous females (per female)}

\item{maxbite}{numeric.  Maximal number of bite per human.}

\item{startFav}{date. First day of the favorable period for diapausing mosquitos.}

\item{endFav}{date. Last day of the favorable period for diapausing mosquitos.}

\item{muErain}{numeric. Additional mortality of eggs due to heavy rain (> 80mm)}

\item{muLrain}{numeric. Additional mortality of larvae due to heavy rain (> 80mm)}

\item{muPrain}{numeric. Additional mortality of pupae due to heavy rain (> 80mm)}

\item{verbose}{logical. Provide additional information on parameters during generation}
}
\value{
Named list of parameters
}
\description{
The functions formats all global parameters required for the model.
}
\details{
Aegypti (with \eqn{temperature\_K_{t}} the daily temperature in kalvin):

egg development function: \eqn{(q1E \times temperature\_K_{t} \times q2E \times e^{(q4E - (1 / temperature\_K_{t}))}) / (1 +  e^{q5E \times (q6E - 1 / temperature\_K_{t})})}

transition function from larva to pupa: \eqn{(q1L \times temperature\_K_{t} \times q2L) \times e^{q3L \times (q4L - 1 / temperature\_K_{t})} / (1 + e^{q5L \times (q6L - 1 / temperature\_K_{t})})}

transition function from pupa to emerging adult: \eqn{(q1P \times temperature\_K_{t} \times q2P) \times e^{q3P \times (q4P - 1 / temperature\_K_{t})} / (1 + e^{q5P \times (q6P- 1 / temperature\_K_{t})})}

function for eggs maturation after blood meal: \eqn{(q1Ag \times temperature\_K_{t} \times q2Ag ) \times e^{q3Ag \times (q4Ag - 1 / temperature\_K_{t})} / (1 + e^{q5Ag \times (q6Ag - 1 / temperature\_K_{t})})}

Albopictus (with \eqn{temperature\_C_{t}} the daily temperature in celsius degree):

egg development function: \eqn{(temperature\_C_{t} - TE) / TDDE}

larvae mortality function: \eqn{ mu1L \times e^{(temperature\_C_{t}-10) \times mu2L}  + mu3L}

transition function from larva to pupa: \eqn{q1L \times temperature\_C_{t}^{2}  + q2L \times temperature\_C_{t}+ q3L}

transition function from pupae to emerging adult: \eqn{q1P \times temperature\_C_{t}^{2}  + q2P \times temperature\_C_{t} + q3P}

function for eggs maturation after blood meal: \eqn{(temperature\_C_{t} - TAG )/ TDDAG}

Both:

Larvae mortality function: \eqn{(mu1L \times e^{(temperature\_C_{t} - 10)\times mu2L} + mu3L)\times (1 + Lm/kL)}, with Lm the number of larvae and kL the carrying capacity (density dependance)

Pupae mortality function: \eqn{mu1P \times e^{(temperature\_C_{t}-10) \times  mu2P}  + mu3P}
}
\examples{
build_gdata()

}
