% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ldata.R
\name{build_ldata}
\alias{build_ldata}
\title{Generate local parameters matrix}
\usage{
build_ldata(
  parcels,
  meteo,
  gdata = NULL,
  vector = "Ae. albopictus (D)",
  virus = "DEN",
  mMov = NULL,
  prev_control = NULL,
  start_date = NULL,
  end_date = NULL,
  nYR_init = 1
)
}
\arguments{
\item{parcels}{data.frame or data.table describing the patches. Required columns: 'ID', 'POP', 'Kfix', 'Kvar', 'STATION' (optional), 'DIFF_ALT' (optional).}

\item{meteo}{data.frame or data.table reporting daily meteorological data. Required columns: 'ID', 'DATE', 'RR', 'TP'.}

\item{gdata}{list of parameters, can be generated using `build_gdata` function.}

\item{vector}{string. Default is "Ae. albopictus (D)".}

\item{virus}{string. Default is "DEN" (dengue virus).}

\item{mMov}{list of two matrices for movement probabilities. Defaults to NULL, which means no movement.}

\item{prev_control}{data.frame or data.table describing preventive control measures. Required columns: 'action', 'loc', 'start', 'end', 'p'.}

\item{start_date}{date in 'Y-m-d' format. Defaults to the earliest date in the meteorological dataset.}

\item{end_date}{date in 'Y-m-d' format. Defaults to the most recent date in the meteorological dataset.}

\item{nYR_init}{numeric. Number of years used to initialize population dynamics. Default is 1.}
}
\value{
matrix of local parameters.
}
\description{
The function computes and formats for each patch the list of local parameters required by the model.
}
