% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mMov.R
\name{build_mMov}
\alias{build_mMov}
\title{Build matrix of contact probabilities}
\usage{
build_mMov(
  SpatVec,
  law = "NGravExp",
  param = NULL,
  p2move = NULL,
  outflow = NULL,
  inflow = NULL,
  verbose = F
)
}
\arguments{
\item{SpatVec}{SpatVector or sf. Spatial vector of polygons representing patches. Polygons must include 'POP' and 'ID' attributes.}

\item{law}{string. Law used to calculate probabilities. Default is "NGravExp". See documentation of `TDLM::run_law` function for law options details.}

\item{param}{numeric. Parameter used to adjust the importance of distance or opportunity associated with the chosen law.
A single value or a vector of several parameter values can be used.
Not necessary for the original radiation law or the uniform law. (see TDLM package for more details)
Default is NULL. If NULL but required the parameter will be estimated by TDLM package.}

\item{p2move}{numeric. Optional. Daily probability to move from the residential patch or daily proportion of residents moving from the residential patch. p2move is required if outflow is not provided. If the outflow is provided, the p2move probability is not used.}

\item{outflow}{numeric. Optional. Average number of person moving out of each patch daily.}

\item{inflow}{numeric. Optional. Average number of person moving in each patch daily.}

\item{verbose}{logical. Display more information during calculation}
}
\value{
List of mobility metrics:

`mi_mixedpop` is the estimated daily number of persons in each administrative unit (staying residents + visitors).

`proba_ij` is a normalized matrix of probabilities for movements of residents from patch i (rows) in patch j in j (columns) (proba_ij). The referent population is the resident population.

`proba_ji` is a normalized matrix of probabilities for origin i (columns) of agents in patch j (rows) (proba_ji). The referent population is the total population during the day including staying residents and visitors.
}
\description{
This function estimate probability matrices of movements/trips between all the polygons of a SpatVector or sf object and return them along with other mobility metrics.
}
