% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prev_control.R
\name{build_prev_control}
\alias{build_prev_control}
\title{Generate the list of preventive control to implement in the simulation}
\usage{
build_prev_control(
  action,
  lon,
  lat,
  start,
  end,
  p,
  SpatVec,
  prev_control = NULL,
  buffer_width = NULL,
  plot_buffer = F
)
}
\arguments{
\item{action}{string. 'K': Source reduction (removal or destruction of breeding sites); 'L': Chemical Larviciding; 'A': Fogging or Area Spraying (targets adult mosquitoes)}

\item{lon}{string. longitude of the control location}

\item{lat}{string. latitude of the control location}

\item{start}{date in '\%Y-\%m-\%d' format. Define the beginning of the implemented measure}

\item{end}{date in '\%Y-\%m-\%d' format. Define the end of the implemented measure}

\item{p}{number between 0 and 1. It is for the "K" action: the proportion of sites daily removed during the action ; for the "A" action: the additional daily mortality of adults due to action and for the "L" action: the additional daily mortality of larvae due to larvicide}

\item{SpatVec}{SpatVector or sf. (Required with buffer) Spatial vector of polygons representing patches. Polygons must include 'ID' attribute including the 'loc'.}

\item{prev_control}{data.frame. (optional) prev_control object to build on (add new measures)}

\item{buffer_width}{integer. (optional) Buffer around the location to implement the measure}

\item{plot_buffer}{logical. Display plot of the selected parcels in case of spatial buffer.}
}
\value{
events data.frame with preventive control measures
}
\description{
Function used to generate the preventive control data.frame.

Note that the distance unit of the buffer width parameter is meters if the CRS is (+proj=longlat), and in map units (typically also meters) if not.
}
\examples{

f <- system.file("shape/SpatVec.shp", package = "arbocartoR")
SpatVec <- terra::vect(f)
build_prev_control(action = "K",
      lon = 1022397, lat = 6321347,
      start = "2022-01-01", end = "2022-01-05",
      p = 0.2,
      SpatVec = SpatVec,
      buffer_width = 100)

}
\keyword{events}
