% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_transitions.R
\name{build_transitions}
\alias{build_transitions}
\title{Write stochatic transitions}
\usage{
build_transitions(gdata)
}
\arguments{
\item{gdata}{list of parameters}
}
\value{
String vector describing transitions
}
\description{
Internal function to build the string vector describing stochastic transitions and counters. This vector is required by SimInf package.
}
\keyword{internal}
