% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{configK}
\alias{configK}
\title{Carrying capacity estimates per land cover type}
\format{
data frame with columns:
\describe{
  \item{ID}{...}
}
}
\source{
mtd
}
\description{
Carrying capacity estimates per land cover type
}
\keyword{data}
