% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_K.R
\name{estim_K}
\alias{estim_K}
\title{Estimate spatial carrying capacities}
\usage{
estim_K(PARCELS_SHAPE, URBAN_ATLAS = NULL)
}
\arguments{
\item{PARCELS_SHAPE}{SpatVector object with polygons of parcels}

\item{URBAN_ATLAS}{SpatVector object or list of SpatVector objects with polygons of soil occupation class by urban atlas}
}
\value{
SpatVector object
}
\description{
Function used to estimate carrying capacity from soil occupation classes
}
\examples{
\dontrun{
PARCELS_SHAPE <- system.file("shape/SpatVec.shp", package = "arbocartoR") \%>\% vect
PARCELS_SHAPE_K_ESTIM <- estim_K(PARCELS_SHAPE, URBAN_ATLAS)
}

}
\keyword{capacity}
\keyword{carrying}
