% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_meteo.R
\name{filter_meteo}
\alias{filter_meteo}
\title{Filter Meteorological Data}
\usage{
filter_meteo(parcels, meteo, TS_sim)
}
\arguments{
\item{parcels}{data.table with parcel information, including a column `ID` or `STATION`.}

\item{meteo}{data.table with meteorological data, including columns `ID` and `DATE`.}

\item{TS_sim}{list containing a vector `time_serie_date` of dates to be retained.}
}
\value{
data.table with filtered meteorological data.
}
\description{
Filters meteorological data to retain only the rows corresponding to parcel IDs present in `parcels` and dates present in `TS_sim$time_serie_date`.
}
\details{
This function ensures that the meteorological data contains only the IDs found in `parcels` and that there are records for all dates in `TS_sim$time_serie_date`.
}
\examples{
# Example usage:
parcels <- data.table(ID = 1:3, STATION = c('A', 'B', 'C'))
meteo <- data.table(ID = c('A', 'B', 'C'), DATE = as.Date('2023-01-01') + 0:2)
TS_sim <- list(time_serie_date = as.Date('2023-01-01') + 0:2)
filter_meteo(parcels, meteo, TS_sim)

}
\keyword{filter}
\keyword{meteo}
