% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo}
\alias{meteo}
\title{Meteorological data}
\format{
data frame with columns:
\describe{
  \item{ID}{unique identifier. If the meteorological data are associated to meteorological stations the unique identifier is the one of the meteorological station (reflecting the 'STATION' column of the parcel dataset), otherwise the identifier is the one of the associated parcel.}
  \item{DATE}{date of the meteorological record}
  \item{RR}{daily precipitation/rainfall (in mm)}
  \item{TP}{daily average temperature (in degree)}


}
}
\source{
meteofrance
}
\description{
Meteorological data
}
\keyword{data}
