% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parcels}
\alias{parcels}
\title{Parcelle data}
\format{
data frame with columns:
}
\source{
iris-ge, urban atlas, CLC
}
\description{
The STATION, DIFF_ALT, and ALT are mandatory in case of using data from meteorological stations and should be removed if the IDs of the meteorological dataset reflect the parcels IDs.
\describe{
  \item{ID}{unique identifier for each parcel}
  \item{NOM_COM}{name of the city associated to the parcel (optional)}
  \item{SURF_HA}{surface in hectare (optional)}
  \item{STATION}{unique identifier of the meteorological station associated to the parcel (optional)}
  \item{DIFF_ALT}{difference between average altitude of the parcel and the altitude of the associated meteorological station (optional)}
  \item{ALT}{average altitude of the parcel (optional)}
  \item{Kfix}{fix carrying capacity (number of larvae in anthropic breeding sites)}
  \item{Kvar}{varying carrying capacity (number of larvae in natural breeding sites)}
  \item{POP}{human population size}

  }
}
\keyword{data}
