% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TS.R
\name{plot_TS}
\alias{plot_TS}
\title{Plot trajectories of the first simulation}
\usage{
plot_TS(traj, stage, parcels_ids = NULL, simulation = 1)
}
\arguments{
\item{traj}{output of a run_arbocartoR simulation}

\item{stage}{String vector. Epidemiological or biological stages to visualize (see colnames of traj).}

\item{parcels_ids}{String vector. Patch ids to visualize (must match with ID column in traj objects).}

\item{simulation}{Numerical vector. Considered simulation (must be <= to the length of traj).}
}
\value{
dygraphs plot
}
\description{
Function used to plot the trajectories of any compartments (median and 95% interval over all the selected parcels and simulation).
}
\examples{

\dontrun{
data(parcels)
data(meteo)

parcels <- parcels[startsWith(ID, "06"),]

traj <- run_arbocartoR(parcels = parcels,
                       vector = "Ae. albopictus",
                       virus = "DEN",
                       meteo = meteo)

plot_TS(traj, stage = c("Em", "Lm", "Pm"), parcels_ids = NULL, simulation = 1)
}

}
\keyword{plot}
\keyword{trajectories}
\keyword{visualization}
