% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_arbocartoR.R
\name{run_arbocartoR}
\alias{run_arbocartoR}
\title{Run a simulation}
\usage{
run_arbocartoR(
  parcels,
  vector = "Ae. albopictus (D)",
  virus = "DEN",
  gdata = NULL,
  ldata = NULL,
  meteo = NULL,
  mMov = NULL,
  start_date = NULL,
  end_date = NULL,
  nYR_init = 1,
  n_sim = 1,
  introduction_pts = NULL,
  prev_control = NULL,
  ref_norm = 100,
  u0 = NULL,
  v0 = NULL,
  initMosq = 1e+05,
  verbose = F
)
}
\arguments{
\item{parcels}{data.frame or data.table describing the patches. Required columns: 'ID': unique identifier, 'POP': population size, 'Kfix': fix carrying capacity for larvae and pupae, 'Kvar': variable carrying capacity for larvae and pupae, 'STATION': meteorological station identifier (optional), 'DIFF_ALT': difference between meteorological station altitude and average altitude of the patch (optional).}

\item{vector}{string. "Ae. albopictus", "Ae. albopictus (D)" or "Ae. aegypti". Default is "Ae. albopictus".}

\item{virus}{string.  "DEN" (dengue), "ZIK" (zika) or "CHI" (chikungunya). Default is "DEN" (dengue) virus.}

\item{gdata}{list of parameters. Can be generated using `build_gdata` function.}

\item{ldata}{matrix of local data}

\item{meteo}{data.frame or data.table reporting the daily meteorological data for each meteorological station. Required columns: 'ID', 'DATE', 'RR': daily precipitation (mm), 'TP': daily mean temperature (degrees)}

\item{mMov}{double matrix}

\item{start_date}{date in '\%Y-\%m-\%d' format}

\item{end_date}{date in '\%Y-\%m-\%d' format}

\item{nYR_init}{numeric. Number of years used to initialize the population dynamics (default is 1)}

\item{n_sim}{integer}

\item{introduction_pts}{data.frame or data.table describing the introduction of individuals. Can be generated by [build_E_random()] function (see function documentation for additional details on the structure.}

\item{prev_control}{data.frame or data.table describing preventive control measure implemented. Required columns: 'action', 'loc', 'start', 'end', 'p' (see details)}

\item{ref_norm}{numeric, rainfall normalization factor / reference value (>0)}

\item{u0}{data.frame describing the initial population stage for each compartment in each parcel. Can be generated by [iniState()] function (see documentation of the function for more details on the structure).}

\item{v0}{data.frame describing the initial population and time-dependant parameters state.  Can be generated by [iniState()] function (see documentation of the function for more details on the structure).}

\item{initMosq}{numeric. Number of eggs in each patch at t0 (default is 100000).

 @details
 Preventive control content:
 'action' column must be strings 'K', 'L' or 'A'. 'K': Source reduction (removal or destruction of breeding sites); 'L': Chemical Larviciding; 'A': Fogging or Area Spraying (targets adult mosquitoes)
 'loc' column must be a parcel id
 'start' is the first day of implementation of the measure
 'end' is the last day of implementation of the measure (the control is implemented every day in between)
 'p' must be a number between 0 and 1. It is for the "K" action: the proportion of sites daily removed during the action ; for the "A" action: the additional daily mortality of adults due to action and for the "L" action: the additional daily mortality of larvae due to larvicide}

\item{verbose}{logical. Provide additional information during process}
}
\value{
data.table
}
\description{
function to run simulations
}
